Add more checks for gizmo interaction
When a gizmo is not visible there is an early return so the checks for
gizmo interaction are never done. If that non-visible gizmo was the
interacted gizmo, it will keep the interaction status and disallow other
gizmos from being interacted with. Add a check in the !Visible check
block and reset interaction status.

Since interaction checking is only done when any gizmo is enabled, if
a gizmo is disabled while it's being interacted with and then destroyed,
no other gizmo can be interacted with. A new check is added in
UpdateInteracting for gizmoInteracting and the left mouse button is not
held. This ensures that as long as there's one active gizmo, interaction
checks are made.